Entities y repositories BD SQL server

+++ Java/toures-b2c-java/src/main/java/co/com/toures/b2c/touresb2cjava/dao/admpro/CityRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.Optional;
 
-public interface CityRepository extends CrudRepository<City, Long> {
+public interface CityRepository extends CrudRepository<City, Integer> {
 
     City findById(int idCity);
 }
